+++ examples/menu.jl
@@ -20,6 +20,7 @@ using REPL.TerminalMenus
 options = [
     "v3kite = include(\"v3kite.jl\")",
     "v3beam = include(\"v3beam.jl\")",
+    "v3beam_replay = include(\"v3beam_replay.jl\")",
     "v3beam_geometry = include(\"v3beam_geometry.jl\")",
     "relax_bridle = include(\"relax_bridle.jl\")",
     "reel_out_v3 = include(\"reel_out_v3.jl\")",