+++ src/V3Kite.jl
@@ -35,7 +35,7 @@ export SymbolicAWEModel, SystemStructure, Logger, SysState
 export load_sys_struct_from_yaml, set_data_path, get_data_path
 export init!, next_step!, update_sys_state!, log!, save_log, load_log, replay
 export PARTICLE_DYNAMICS, RIGID_DYNAMICS, AeroDirect, ContinuousAero
-export KernelBackend, MonolithBackend
+export AeroPressure, KernelBackend, MonolithBackend
 export Settings
 export SymbolicAWEModels
 export record, replay